chore: Moved settings related queries to new structure
- We provide a `staleTime` & `gcTime` for the `useLatestRelease` hook as we don't want to ping the GitHub api too many times
- Holding off on mutations from now
- Allows us to remove `@/constants/QueryKeys.ts`
